Fix configs
